refact to use FSD architecture
