--- atoms/medication.atom.ts
@@ -1,5 +1,0 @@
-import { atom } from 'jotai'
-import { Medication, ListMedicationAtom } from '@types'
-
-export const medicationsAtom = atom<ListMedicationAtom[]>([])
-export const selectedMedicationAtom = atom<Medication>()
