--- components/Medications/MedicationInfo/styles.ts
@@ -1,16 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export default StyleSheet.create({
-  container: {
-    marginVertical: 8,
-  },
-  coverImage: {
-    borderTopLeftRadius: 16,
-    borderTopRightRadius: 16,
-    borderBottomLeftRadius: 0,
-    borderBottomRightRadius: 0,
-  },
-  infoContainer: {
-    marginVertical: 16,
-  }
-})
