--- components/Medications/MedicationList/styles.ts
@@ -1,10 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-export default StyleSheet.create({
-  listcontainer: {
-    marginTop: 36
-  },
-  cardContainer: {
-    marginBottom: 12,
-  },
-})
