--- components/common/Dialog/styles.ts
@@ -1,10 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-export default StyleSheet.create({
-  title: {
-    textAlign: 'center'
-  },
-  actionsContainer: {
-    marginTop: 24 
-  }
-})
