--- components/common/TextInput/styles.ts
@@ -1,7 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-export default StyleSheet.create({
-  errorMessage: {
-    marginLeft: 16,
-  },
-})
