+++ features/medicines/components/MedicationCard/styles.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  cardContainer: {
+    width: "100%",
+  },
+  image: {
+    width: 80,
+    height: 80,
+    borderTopRightRadius: 12,
+    borderBottomRightRadius: 12,
+  },
+});
