+++ features/medicines/components/MedicationList/styles.ts
@@ -1,0 +1,10 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  listcontainer: {
+    marginTop: 36,
+  },
+  cardContainer: {
+    marginBottom: 12,
+  },
+});
