+++ features/medicines/screens/medicationRegister/styles.ts
