--- hooks/common/useDebounce.ts
@@ -1,17 +1,0 @@
-import { useState, useEffect } from 'react'
-
-function useDebounce<T>(value: T, delay: number = 500): T {
-  const [debouncedValue, setDebouncedValue] = useState<T>(value);
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setDebouncedValue(value)
-    }, delay)
-
-    return () => clearTimeout(timer)
-  }, [value, delay])
-
-  return debouncedValue;
-}
-
-export default useDebounce
