+++ shared/components/common/Dialog/styles.ts
@@ -1,0 +1,10 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  title: {
+    textAlign: "center",
+  },
+  actionsContainer: {
+    marginTop: 24,
+  },
+});
