+++ shared/components/common/TextInput/styles.ts
@@ -1,0 +1,7 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  errorMessage: {
+    marginLeft: 16,
+  },
+});
