+++ tsconfig.json
@@ -9,7 +9,6 @@
       ],
       "@types": ["./types/index.d.ts"],
       "@atoms/*": ["./atoms/*"],
-      "@schemas/*": ["./schemas/*"],
     }
   },
   "include": [
